Separated Products from product card
