+++ src/products.js
@@ -1,0 +1,54 @@
+const products = [
+    {
+      title: "Aluminum Sliding Windows",
+      desc: "Premium aluminum sliding windows with double glazing and thermal break technology.",
+      img:"/images/aluminium.jpg", 
+      // price: "€185.00",
+      // unit: "per window",
+      category: "Aluminum Windows",
+      rating: 4.9,
+      stock: true,
+    },
+    {
+      title: "Aluminum Entry Doors",
+      desc: "Modern aluminum entry doors with security features and energy efficiency.",
+      img: "/images/alum_doors.jpg",
+      // price: "€320.00",
+      // unit: "per door",
+      category: "Aluminum Doors",
+      rating: 4.8,
+      stock: true,
+    },
+    {
+      title: "Concrete Blocks",
+      desc: " High-performance insulation material designed for modern construction. Engineered with sustainability and efficiency in mind, our product provides superior thermal performance, reduces energy costs, and contributes to a healthier indoor environment..",
+      img: "/images/block1.jpg",
+      // price: "€165.00",
+      // unit: "per window",
+      category: "Concrete Blocks",
+      rating: 4.7,
+      stock: true,
+    },
+    {
+      title: "Tiles",
+      desc: " High-performance insulation material designed for modern construction. Engineered with sustainability and efficiency in mind, our product provides superior thermal performance, reduces energy costs, and contributes to a healthier indoor environment..",
+      img: "/images/tiles2.jpg",
+      // price: "€165.00",
+      // unit: "per window",
+      category: "tiles",
+      rating: 4.7,
+      stock: true,
+    },
+    {
+      title: "Tandoor Stones",
+      desc: " High-performance insulation material designed for modern construction. Engineered with sustainability and efficiency in mind, our product provides superior thermal performance, reduces energy costs, and contributes to a healthier indoor environment..",
+      img: "/images/tandoor_stone.jpg",
+      // price: "€165.00",
+      // unit: "per window",
+      category: "tiles",
+      rating: 4.7,
+      stock: true,
+    },
+  ];
+
+  export default products
